Move 'visitable' to private folder
